Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,20 +1,19 @@
 class Solution {
-    public int mySqrt(long x) {
-        
-        long left = 1;
-        long right = x + 1;
+    public int mySqrt(int x) {
+        int left = 1;
+        int right = x + 1;
 
         while(left < right) {
-            long mid = left + (right-left)/2;
+            int mid = left + (right - left)/2;
 
             if(mid*mid <= x) {
-                left = mid+1;
+                left = mid + 1;
             } else {
                 right = mid;
             }
-        } 
+        }
 
-        return (int)left - 1;
+        return left - 1;
         
     }
 }
